Begun the move to netstandard.  Removed old projects that aren't necessary anymore and renamed the PCL project to Universal.
