+++ .gitignore
@@ -215,3 +215,4 @@ pip-log.txt
 .mr.developer.cfg
 /Lorem.NET/project.lock.json
 /Lorem.DNX.NET/project.lock.json
+/docs/_build
