--- Lorem.DNX.NET.sln
@@ -1,70 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25123.0
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{438E4420-54B2-45AC-B8C8-010BE97875F3}"
-	ProjectSection(SolutionItems) = preProject
-		appveyor.yml = appveyor.yml
-		global.json = global.json
-		license.md = license.md
-		nuget.config = nuget.config
-		Lorem.NET\nuget.exe = Lorem.NET\nuget.exe
-		postbuild.ps1 = postbuild.ps1
-		prebuild.ps1 = prebuild.ps1
-		readme.md = readme.md
-	EndProjectSection
-EndProject
-Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Lorem.DNX.NET", "Lorem.DNX.NET\Lorem.DNX.NET.xproj", "{FF8ED9D9-7770-4B09-A4AB-54FD868FFE65}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Lorem.Universal.NET", "Lorem.Universal.NET\Lorem.Universal.NET.csproj", "{E23C5AF5-8DD7-410E-B145-AFE89C7293AB}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Debug|ARM = Debug|ARM
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|Any CPU = Release|Any CPU
-		Release|ARM = Release|ARM
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{FF8ED9D9-7770-4B09-A4AB-54FD868FFE65}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{FF8ED9D9-7770-4B09-A4AB-54FD868FFE65}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{FF8ED9D9-7770-4B09-A4AB-54FD868FFE65}.Debug|ARM.ActiveCfg = Debug|Any CPU
-		{FF8ED9D9-7770-4B09-A4AB-54FD868FFE65}.Debug|ARM.Build.0 = Debug|Any CPU
-		{FF8ED9D9-7770-4B09-A4AB-54FD868FFE65}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{FF8ED9D9-7770-4B09-A4AB-54FD868FFE65}.Debug|x64.Build.0 = Debug|Any CPU
-		{FF8ED9D9-7770-4B09-A4AB-54FD868FFE65}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{FF8ED9D9-7770-4B09-A4AB-54FD868FFE65}.Debug|x86.Build.0 = Debug|Any CPU
-		{FF8ED9D9-7770-4B09-A4AB-54FD868FFE65}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{FF8ED9D9-7770-4B09-A4AB-54FD868FFE65}.Release|Any CPU.Build.0 = Release|Any CPU
-		{FF8ED9D9-7770-4B09-A4AB-54FD868FFE65}.Release|ARM.ActiveCfg = Release|Any CPU
-		{FF8ED9D9-7770-4B09-A4AB-54FD868FFE65}.Release|ARM.Build.0 = Release|Any CPU
-		{FF8ED9D9-7770-4B09-A4AB-54FD868FFE65}.Release|x64.ActiveCfg = Release|Any CPU
-		{FF8ED9D9-7770-4B09-A4AB-54FD868FFE65}.Release|x64.Build.0 = Release|Any CPU
-		{FF8ED9D9-7770-4B09-A4AB-54FD868FFE65}.Release|x86.ActiveCfg = Release|Any CPU
-		{FF8ED9D9-7770-4B09-A4AB-54FD868FFE65}.Release|x86.Build.0 = Release|Any CPU
-		{E23C5AF5-8DD7-410E-B145-AFE89C7293AB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E23C5AF5-8DD7-410E-B145-AFE89C7293AB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E23C5AF5-8DD7-410E-B145-AFE89C7293AB}.Debug|ARM.ActiveCfg = Debug|ARM
-		{E23C5AF5-8DD7-410E-B145-AFE89C7293AB}.Debug|ARM.Build.0 = Debug|ARM
-		{E23C5AF5-8DD7-410E-B145-AFE89C7293AB}.Debug|x64.ActiveCfg = Debug|x64
-		{E23C5AF5-8DD7-410E-B145-AFE89C7293AB}.Debug|x64.Build.0 = Debug|x64
-		{E23C5AF5-8DD7-410E-B145-AFE89C7293AB}.Debug|x86.ActiveCfg = Debug|x86
-		{E23C5AF5-8DD7-410E-B145-AFE89C7293AB}.Debug|x86.Build.0 = Debug|x86
-		{E23C5AF5-8DD7-410E-B145-AFE89C7293AB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E23C5AF5-8DD7-410E-B145-AFE89C7293AB}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E23C5AF5-8DD7-410E-B145-AFE89C7293AB}.Release|ARM.ActiveCfg = Release|ARM
-		{E23C5AF5-8DD7-410E-B145-AFE89C7293AB}.Release|ARM.Build.0 = Release|ARM
-		{E23C5AF5-8DD7-410E-B145-AFE89C7293AB}.Release|x64.ActiveCfg = Release|x64
-		{E23C5AF5-8DD7-410E-B145-AFE89C7293AB}.Release|x64.Build.0 = Release|x64
-		{E23C5AF5-8DD7-410E-B145-AFE89C7293AB}.Release|x86.ActiveCfg = Release|x86
-		{E23C5AF5-8DD7-410E-B145-AFE89C7293AB}.Release|x86.Build.0 = Release|x86
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
