+++ Lorem.Universal.NET.sln
@@ -1,0 +1,40 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.26228.4
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{438E4420-54B2-45AC-B8C8-010BE97875F3}"
+	ProjectSection(SolutionItems) = preProject
+		appveyor.yml = appveyor.yml
+		license.md = license.md
+		nuget.config = nuget.config
+		Lorem.NET\nuget.exe = Lorem.NET\nuget.exe
+		postbuild.ps1 = postbuild.ps1
+		prebuild.ps1 = prebuild.ps1
+		readme.md = readme.md
+	EndProjectSection
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "docs", "docs", "{F7DA0599-C83D-4765-8AD0-E1F756032F05}"
+	ProjectSection(SolutionItems) = preProject
+		docs\conf.py = docs\conf.py
+		docs\index.rst = docs\index.rst
+		docs\lorem-class.rst = docs\lorem-class.rst
+	EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Lorem.Universal.NET", "Lorem.Universal.NET\Lorem.Universal.NET.csproj", "{342DEBEE-436C-4E10-A230-0172795E3CBF}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{342DEBEE-436C-4E10-A230-0172795E3CBF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{342DEBEE-436C-4E10-A230-0172795E3CBF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{342DEBEE-436C-4E10-A230-0172795E3CBF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{342DEBEE-436C-4E10-A230-0172795E3CBF}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
