--- Lorem.Universal.NET/project.json
@@ -1,23 +1,0 @@
-{
-  "version": "1.0.0",
-  "description": "A .NET library for all things random!  Intended for Windows 10 Universal Projects.",
-  "authors": [ "dochoffiday", "Tony Richards (trichards57)", "xfischer" ],
-  "tags": [ "Lorem", "Ipsem", "random", "UWP", "rng" ],
-  "projectUrl": "https://github.com/trichards57/Lorem.DNX.NET",
-  "licenseUrl": "https://github.com/trichards57/Lorem.DNX.NET/blob/master/license.md",
-
-  "dependencies": {
-    "Microsoft.NETCore.UniversalWindowsPlatform": "5.0.0"
-  },
-  "frameworks": {
-    "uap10.0": { }
-  },
-  "runtimes": {
-    "win10-arm": { },
-    "win10-arm-aot": { },
-    "win10-x86": { },
-    "win10-x86-aot": { },
-    "win10-x64": { },
-    "win10-x64-aot": { }
-  }
-}
