--- nuget.config
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<configuration>
-  <packageSources>
-    <!--To inherit the global NuGet package sources remove the <clear/> line be -->
-    <clear />
-    <add key="dotnet.cli" value="https://dotnet.myget.org/F/dotnet-cli/api/v3/index.json" />
-    <add key="dotnet.core" value="https://dotnet.myget.org/F/dotnet-core/api/v3/index.json" />
-    <add key="RC2" value="https://www.myget.org/F/aspnetvnext/api/v3/index.json" />
-    <add key="xunit" value="https://www.myget.org/F/xunit/" />
-    <add key="api.nuget.org" value="https://api.nuget.org/v3/index.json" />
-    <add key="aspnet-contrib" value="https://www.myget.org/F/aspnet-contrib/api/v3/index.json" />
-  </packageSources>
-  <disabledPackageSources>
-    <add key="Microsoft and .NET" value="true" />
-    <add key="PostSharp" value="true" />
-  </disabledPackageSources>
-</configuration>
